bot/requirements.txt
peewee==3.16.2
Pillow==10.0.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
python-telegram-bot==13.15
pytz==2023.3
requests==2.22.0